requirements.txt
flask
flask-cors
flask_restx
flask-security-too
flask-mongoengine
pytest-flask
pytest-cov
bcrypt
musicbrainzngs
faker
python-resize-image
coverage
